backend/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
PyGithub
sqlalchemy
psycopg2-binary